ai-service/requirements.txt
fastapi
uvicorn
pytesseract
pdfplumber
pillow
requests
spacy
python-dotenv
